Change of logic -> change of name.
